Updates to be more robust to switches in underlying dataset with new ReactiveFacileDataStore
I think the problem was:

When a new datastore was slipped into the rfds via datasetSelect, I was ras running into
problems with the sample filter because the input$values in the categoricalCovariateSelect
and levels were updated on the server through updateSelectizeInput, but reactives responding
during the same round weren't getting these new values yet.

Trying to mitigate that.

WOW, that was fun.