Fixed state double-dipping in assaySelect.
There is an antipattern that needs to be avoided, which is to have
have the same reactive block depend on both an `input$something`
and the internal `state$something` equivalent.

This block:

```
assay_info <- eventReactive(list(input$assay, state$assay_info), {
  # ...
})
```

Was bad because both the input$ and state$ were colliding with each other.
It was split into two blocks:

```
observeEvent(input$assay, {
  # which does some junk here and updates
  # state$assay_info, if necessary
})

assay_info <- reactive(state$info)
```

That's the gist of it