Giving up on being fancy with categoricalSampleCovariateSelect and mutex
Currently ignoreNULL is set to true on these things, and this doesn't
"flush" the empty select and release the levels of a select box. This bites
you when trying to have one selectlevel input act as the `.exclude` for the
other so you can have mutually exclusive selectLevels from the same
covariate.