Ensures current samples are from current backing datastore
Switching the underlying datastore can cause the current active_samples(rfds) to
be set to the old samples when the datastore has been updated. See this issue:

https://github.com/denalitherapeutics/DenaliEyrie/issues/2

This is a hard bug to reproduce, and happens in weird situations when switching
datasets is the first thing done in The Eyrie ... when the app restarts, that
isn't even reproducible. I must be missing something, but can't figure it out.

Including the `test_sample_subset(samples, rfds)` bounces out of the
fetch_sample_covarites in a **seemingly** safe way for the app to try again
without crashing.